scripts/requirements.txt
pandas==1.5.3
numpy==1.24.2
kafka-python==2.0.2
confluent-kafka==2.3.0
requests==2.31.0
pyspark==3.5.0